Rename DATA PREPROCESSING.py to Predicting Biomarkers in Alzheimer's disease patients using a Multimodel Convoluted Neural Network.py
Alzheimer's disease is one of the most common neurodegenerative diseases that is affecting millions of people. Dementia is a common symptom of Alzheimer's and it is usually diagnosed using clinical tests and brain pathology. Currently, the diagnosis is done with the help of MRI scans. The identification of  Biomarkers was developed for improving their diagnosis and helping accelerate the development of new therapies. Thus given the recent breakthrough in Deep Learning as it provides a quicker and more accurate diagnosis which provides a solution to prevent medical errors in this disease. We constructed a Convolutional neural network to perform Alzheimer's Disease Image classification, The classification results of the proposed model show results of 98% accuracy.

To initiate the CNN, the data must be acquired and preprocessed for effective model training 
The data used in this paper is from the ADNI Dataset (Alzheimer's Disease Neuroimaging Initiative), It is a project launched in 2003 as a public-private partnership. The goal of the ADNI study is to detect AD at the earliest possible stage and support advances in intervention, prevention, and treatment through new diagnostic methods. The data comprised 1059 MRI scans containing age groups from 55-90 years old. These MRI scans are in the form of 2D axial images in NIFTII format.

All the data after the collection was preprocessed through 4 classifiers to make sure that the MRI images can be efficiently tasked for image classification and model training. These classifiers are 1. Skull stripping
                       2. Bias Correction
                       3. Otsu threshold
                       4. Support Vector Machines

From the Model training the results indicate that the model's accuracy is showing 98% which means measures the number of correct predictions made divided by the total number of predictions made. The loss calculated shows as  0.1765 this means the number indicating how bad the model's prediction was on a single example.